--- presentation/src/main/java/com/madrid/presentation/screens/searchScreen/SearchInteractionLisener.kt
@@ -1,5 +1,0 @@
-package com.madrid.presentation.screens.searchScreen
-
-interface SearchInteractionListener {
-    fun onSearchQuerySubmitted(query: String,viewModel:SearchViewModel)
-}
