+++ presentation/src/main/java/com/madrid/presentation/screens/searchScreen/viewModel/interactionListener/SearchInteractionLisener.kt
@@ -1,0 +1,7 @@
+package com.madrid.presentation.screens.searchScreen.viewModel.interactionListener
+
+import com.madrid.presentation.screens.searchScreen.viewModel.SearchViewModel
+
+interface SearchInteractionListener {
+    fun onSearchQuerySubmitted(query: String,viewModel: SearchViewModel)
+}
